+++ src/services/Application/submitApplication.ts
@@ -1,9 +1,9 @@
-import { axiosPrivate } from "@/src/services/api/AxiosPrivate";
+import { axiosPrivate } from "@/src/services/api/customAxios";
 import { ApplicationData } from "@/src/types/models/ApplicationData";
 
 export const submitApplication = async (
   data: ApplicationData
 ): Promise<number> => {
-  const response = await axiosPrivate.post("/api/v1/application", data)
+  const response = await axiosPrivate.post("/api/v1/application", data);
   return response.data.applicationId;
 };
