+++ src/services/Application/uploadImage.ts
@@ -1,4 +1,4 @@
-import { axiosPrivate } from "@/src/services/api/AxiosPrivate";
+import { axiosPrivate } from "@/src/services/api/customAxios";
 import { File } from "@/src/types/File";
 
 export const uploadImage = async (file: File): Promise<string> => {
